error handling working on User (so far)

+++ src/main/java/fr/chatop/api/config/JwtTokenFilter.java
@@ -66,7 +66,7 @@ public class JwtTokenFilter extends OncePerRequestFilter {
 		User userDetails = new User();
 		String[] jwtSubject = jwtUtil.getSubject(token).split(",");
 
-		userDetails.setId(Integer.parseInt(jwtSubject[0]));
+		userDetails.setId(Long.parseLong(jwtSubject[0]));
 		userDetails.setEmail(jwtSubject[1]);
 
 		return userDetails;
